+++ neon_utils/skills/neon_skill.py
@@ -51,13 +51,8 @@ from neon_utils.message_utils import dig_for_message, resolve_message, get_messa
 from neon_utils.cache_utils import LRUCache
 from neon_utils.file_utils import resolve_neon_resource_file
 from neon_utils.user_utils import get_user_prefs
+from neon_utils.hana_utils import request_backend, ServerException
 from ovos_workshop.skills.ovos import OVOSSkill
-
-try:
-    from neon_utils.mq_utils import send_mq_request
-except ImportError:
-    LOG.warning("MQ Dependencies not installed")
-    send_mq_request = None
 
 try:
     from ovos_plugin_manager.language import OVOSLangDetectionFactory, \
@@ -74,6 +69,7 @@ SPEED_MODE_EXTENSION_TIME = {
 }
 DEFAULT_SPEED_MODE = "thoughtful"
 CACHE_TIME_OFFSET = 24*60*60  # seconds in 24 hours
+
 
 @deprecated("deprecated without replacement, skill settings no longer shipped in skill folder", "2.0.0")
 def save_settings(skill_dir, skill_settings):
@@ -92,6 +88,7 @@ def save_settings(skill_dir, skill_settings):
         LOG.exception(f'error saving skill settings to {settings_path}')
     else:
         LOG.info(f'Skill settings successfully saved to {settings_path}')
+
 
 class NeonSkill(OVOSSkill):
     def __init__(self, name=None, bus=None, **kwargs):
@@ -379,15 +376,18 @@ class NeonSkill(OVOSSkill):
         if not email_addr and message:
             email_addr = get_user_prefs(message)["user"].get("email")
 
-        if email_addr and send_mq_request:
+        if email_addr:
             LOG.info("Send email via Neon Server")
             request_data = {"recipient": email_addr,
                             "subject": title,
                             "body": body,
                             "attachments": attachments}
-            data = send_mq_request("/neon_emails", request_data,
-                                   "neon_emails_input")
-            return data.get("success")
+            try:
+                request_backend("/email", request_data)
+                return True
+            except ServerException as e:
+                LOG.error(e)
+                return False
         else:
             LOG.warning("Attempting to send email via Mycroft Backend")
             super().send_email(title, body)
@@ -893,22 +893,3 @@ class NeonSkill(OVOSSkill):
         """
         log_deprecation("Use `load_dialog_files`", "2.0.0")
         self.load_dialog_files(root_directory)
-
-    def add_event(self, name: str, handler: callable,
-                  handler_info: Optional[str] = None, once: bool = False,
-                  speak_errors: bool = True):
-        # TODO: Remove with ovos-workshop==0.0.13
-        try:
-            # Patching FakeBus compat. with MessageBusClient
-            if hasattr(self.bus, "ee"):
-                emitter = self.bus.ee
-            else:
-                emitter = self.bus.emitter
-            if handler_info == "mycroft.skill.handler" and \
-                    emitter.listeners(name):
-                LOG.warning(f"Not re-registering intent handler {name}")
-                return
-        except Exception as e:
-            LOG.exception(e)
-        OVOSSkill.add_event(self, name, handler, handler_info, once,
-                            speak_errors)